jwt is complete

+++ src/main/java/com/tss/loanEmiSchedular/entity/Emi.java
@@ -5,6 +5,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -19,11 +20,11 @@ public class Emi extends BaseEntity{
 
     private Integer installmentNumber;
 
-    private Double principal;
-    private Double interest;
+    private BigDecimal principal;
+    private BigDecimal interest;
 
-    private Double penaltyAmount;
-    private Double totalPaidAmount;
+    private BigDecimal penaltyAmount;
+    private BigDecimal totalPaidAmount;
 
     private Boolean isFullyPaid=false;
